modularized eleventy config modules
