+++ lib/plugins/navigation.js
@@ -1,0 +1,5 @@
+import eleventyNavigationPlugin from "@11ty/eleventy-navigation";
+
+export default function (eleventyConfig){
+    eleventyConfig.addPlugin(eleventyNavigationPlugin)
+}
