+++ src/index.md
@@ -1,6 +1,9 @@
 ---
-
+needsHighlight: true
 ---
 # Welcome to the Vexflow Examples!
 
 Click on an example to start learning!
+
+```js
+console.log("hello, welcome to our site!)
